Refactor: Controller Service 구조 변경
